Update imports & move classes out of interactor module.
