+++ app-features/following/build.gradle.kts
@@ -8,7 +8,6 @@ android {
 
 dependencies {
     implementation(projects.appCommon.compose)
-    implementation(projects.shared.interactors)
     implementation(projects.shared.domain.showDetails.api)
     implementation(projects.shared.domain.showCommon.api)
 }
