+++ app-features/following/src/main/kotlin/com/thomaskioko/tvmaniac/following/FollowingContract.kt
@@ -1,4 +1,4 @@
-package com.thomaskioko.tvmaniac.interactors
+package com.thomaskioko.tvmaniac.following
 
 import com.thomaskioko.tvmaniac.shared.core.store.Action
 import com.thomaskioko.tvmaniac.shared.core.store.Effect
