+++ app-features/following/src/main/kotlin/com/thomaskioko/tvmaniac/following/FollowingListScreen.kt
@@ -26,8 +26,6 @@ import com.thomaskioko.tvmaniac.compose.components.LazyGridItems
 import com.thomaskioko.tvmaniac.compose.components.NetworkImageComposable
 import com.thomaskioko.tvmaniac.compose.components.SwipeDismissSnackbar
 import com.thomaskioko.tvmaniac.compose.rememberFlowWithLifecycle
-import com.thomaskioko.tvmaniac.interactors.WatchlistEffect
-import com.thomaskioko.tvmaniac.interactors.WatchlistState
 import com.thomaskioko.tvmaniac.resources.R
 import kotlinx.coroutines.InternalCoroutinesApi
