+++ app-features/following/src/main/kotlin/com/thomaskioko/tvmaniac/following/ObserveFollowingInteractor.kt
@@ -1,4 +1,4 @@
-package com.thomaskioko.tvmaniac.interactors
+package com.thomaskioko.tvmaniac.following
 
 import com.thomaskioko.tvmaniac.datasource.cache.Show
 import com.thomaskioko.tvmaniac.details.api.repository.TvShowsRepository
