+++ app/src/main/kotlin/com/thomaskioko/tvmaniac/injection/InteractorsModule.kt
@@ -8,16 +8,16 @@ import com.thomaskioko.tvmaniac.discover.api.ObserveDiscoverShowsInteractor
 import com.thomaskioko.tvmaniac.discover.api.repository.DiscoverRepository
 import com.thomaskioko.tvmaniac.episodes.api.EpisodeRepository
 import com.thomaskioko.tvmaniac.episodes.api.ObserveEpisodesInteractor
+import com.thomaskioko.tvmaniac.following.ObserveFollowingInteractor
 import com.thomaskioko.tvmaniac.genre.api.GenreRepository
 import com.thomaskioko.tvmaniac.genre.api.GetGenresInteractor
-import com.thomaskioko.tvmaniac.interactors.ObserveFollowingInteractor
-import com.thomaskioko.tvmaniac.interactors.ObserveShowsByCategoryInteractor
 import com.thomaskioko.tvmaniac.lastairepisodes.api.LastAirEpisodeRepository
 import com.thomaskioko.tvmaniac.lastairepisodes.api.ObserveAirEpisodesInteractor
 import com.thomaskioko.tvmaniac.seasonepisodes.api.ObserveSeasonWithEpisodesInteractor
 import com.thomaskioko.tvmaniac.seasonepisodes.api.SeasonWithEpisodesRepository
 import com.thomaskioko.tvmaniac.seasons.api.SeasonsRepository
 import com.thomaskioko.tvmaniac.seasons.api.interactor.ObserveSeasonsInteractor
+import com.thomaskioko.tvmaniac.show_grid.domain.ObserveShowsByCategoryInteractor
 import com.thomaskioko.tvmaniac.similar.api.ObserveSimilarShowsInteractor
 import com.thomaskioko.tvmaniac.similar.api.SimilarShowsRepository
 import dagger.Module
