+++ shared/domain/similar/implementation/build.gradle.kts
@@ -24,4 +24,13 @@ dependencies {
     commonTestImplementation(libs.testing.kotest.assertions)
     commonTestImplementation(libs.testing.mockk.common)
     commonTestImplementation(libs.testing.coroutines.test)
+
+    val coroutineCore = libs.kotlin.coroutines.core.get()
+
+    @Suppress("UnstableApiUsage")
+    iosMainImplementation("${coroutineCore.module.group}:${coroutineCore.module.name}:${coroutineCore.versionConstraint.displayName}") {
+        version {
+            strictly(libs.versions.coroutines.native.get())
+        }
+    }
 }
